pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sa1rb"
version = "0.1.0"
authors = [
    { name = "Richard Bergmair", email = "r.bergmair.gh@rb.uref.email" }
]
description = "Richard's Submission to SerpApi's Code Challenge"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["sa1rb"]

dependencies = [
    "beautifulsoup4>=4.13.4",
    "mini-racer>=0.12.4",
    "PyYAML>=6.0.2",
    "pytest>=8.3.5"
]

[tool.setuptools.packages]
find = { where = ["src"] }  # look for packages inside src/
